0.2.0 / Commands and Testing
Major update!

I worked hard for the past 2 weeks to add a better command system to the Dryad engine. Now, you can create and manage any command with a few easy implementations, and includes the ability to add a new command by adding the java file or files that make it, anywhere in the project.

I also added a Testing enviorment since i couldnt make neovim work with JUnit test and i wanted to spare myself the headache, so i made this testing enviorment that just repurposes the same code from the main project. It took me about 30 minutes to get it up and running, as you can probably tell by how most variables and method names are the same as the CommandManager's.

This update brings:
* Better argument partitioning for commands
* Better partition selection for command autocomplete
* Base test framework
* Better command mangement and creation
* 3 new commands (old ones deprecated)
* Refactorized project structure
* Started documentation on the whole project

Known issues:
- module-info.java missing due to some issues with libraryes that i dont know how to fix yet
- most documnetation is missing
- command autocomplete missing
- command multiline missing
- command pipes missing
- same for tests
- the sdl window pops up even if i set it to be disabled in the frontend env (i know what the problem is)
- the file command is not completed
- theres a bunch of planned commands not implemented.
- all deprecated commands except help are missing in the new version

More info:
* i am planning on switching from SDL to OpenGL, due to my inexperience with sdl, and because it is horrible to work with, since the libsdl4j documentation is as vague as your mom describing where the milk is in the fridge.
* i also added my own testing system, and its for good reason. i did not add it because i hate JUnit Test. Quite the contrary actually. But i couldnt get nvim to work with unit tests, and because i wanted less of a headache, i just made my own in like 30 minutes, by repurposing code from the main project. i am also planning on adding hotreloading for the  testing enviorment, aka, you will be able to recompile the test enviorment from withing the test enviorment, making it super easy to make quick changes to any test. Its not JUnit test, but its not shit.
* as for commands, im planning on adding all the commands that were previously present in the depricated command pallete, but for now, all of them except for "help" (that now works WAY differently) are missing.